Split types/interfaces into separate files
