+++ src/lib/types/lightapi.ts
@@ -1,0 +1,11 @@
+export interface LightAPIBalanceRow {
+	currency: string;
+	contract: string;
+	amount: string;
+	decimals: string;
+}
+
+export interface LightAPIBalanceResponse {
+	account_name: string;
+	balances: LightAPIBalanceRow[];
+}
